requirements.txt
streamlit
langchain>=0.3.0
langchain-community
langchain-groq
langchain-huggingface
langchain-chroma
langchain-experimental
langchain-text-splitters
python-dotenv
pypdf
flashrank
sentence-transformers
pydantic>=2.0.0
pysqlite3-binary